# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/job-offers": {
      "post": {
        "operationId": "JobOffersController_create",
        "summary": "Crea o aggiorna un annuncio di lavoro",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobOfferDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOfferDto"
                }
              }
            }
          }
        },
        "tags": [
          "job-offers"
        ]
      },
      "get": {
        "operationId": "JobOffersController_findAllNew",
        "summary": "Recupera la lista dei nuovi annunci di lavoro non ancora catalogati (stato NEW, esclusi i DISQUALIFIED dall AI)",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "integer"
            }
          },
          {
            "name": "freshness",
            "required": false,
            "in": "query",
            "description": "Filtra per freschezza dell annuncio (HOT, RECENT, AGING, OLD)",
            "schema": {
              "enum": [
                "HOT",
                "RECENT",
                "AGING",
                "OLD"
              ],
              "type": "string"
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "description": "Filtra per priorità AI: HIGH, MEDIUM, LOW, DISQUALIFIED",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minDesireScore",
            "required": false,
            "in": "query",
            "description": "Filtra per punteggio minimo aderenza ai desiderata (0-100)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minCompetenceScore",
            "required": false,
            "in": "query",
            "description": "Filtra per punteggio minimo competenza tecnica (0-100)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minScore",
            "required": false,
            "in": "query",
            "description": "Filtra per punteggio globale minimo (0-100)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "remoteType",
            "required": false,
            "in": "query",
            "description": "Filtra per modalità di lavoro (REMOTE, HYBRID, ON_SITE)",
            "schema": {
              "enum": [
                "REMOTE",
                "HYBRID",
                "ON_SITE",
                "UNSPECIFIED"
              ],
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Cerca per testo nel titolo o nel nome azienda",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "required": false,
            "in": "query",
            "description": "Filtra per fonte dell annuncio (ARBEITNOW, REMOTIVE)",
            "schema": {
              "enum": [
                "ARBEITNOW",
                "REMOTIVE",
                "JOBICY",
                "WE_WORK_REMOTELY"
              ],
              "type": "string"
            }
          },
          {
            "name": "evaluatorModel",
            "required": false,
            "in": "query",
            "description": "Filtra per modello AI utilizzato per l analisi (es. GEMINI_3_1_FLASH_LITE, GEMMA_4_12B)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filtra per stato dell annuncio (es. NEW, SAVED, APPLIED, INTERVIEWING, REJECTED, ARCHIVED)",
            "schema": {
              "enum": [
                "NEW",
                "SAVED",
                "APPLIED",
                "SCREENING",
                "INTERVIEWING",
                "OFFER",
                "ACCEPTED",
                "REJECTED",
                "ARCHIVED"
              ],
              "type": "string"
            }
          },
          {
            "name": "includePending",
            "required": false,
            "in": "query",
            "description": "Includi anche gli annunci non ancora valutati dall AI (default: false)",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/JobOfferDto"
                          }
                        },
                        "meta": {
                          "$ref": "#/components/schemas/PaginatedMetaDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "job-offers"
        ]
      }
    },
    "/job-offers/active": {
      "get": {
        "operationId": "JobOffersController_findActive",
        "summary": "Recupera TUTTI gli annunci attivi in gestione senza paginazione (SAVED, APPLIED, SCREENING, INTERVIEWING, OFFER)",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "integer"
            }
          },
          {
            "name": "freshness",
            "required": false,
            "in": "query",
            "description": "Filtra per freschezza dell annuncio (HOT, RECENT, AGING, OLD)",
            "schema": {
              "enum": [
                "HOT",
                "RECENT",
                "AGING",
                "OLD"
              ],
              "type": "string"
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "description": "Filtra per priorità AI: HIGH, MEDIUM, LOW, DISQUALIFIED",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minDesireScore",
            "required": false,
            "in": "query",
            "description": "Filtra per punteggio minimo aderenza ai desiderata (0-100)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minCompetenceScore",
            "required": false,
            "in": "query",
            "description": "Filtra per punteggio minimo competenza tecnica (0-100)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minScore",
            "required": false,
            "in": "query",
            "description": "Filtra per punteggio globale minimo (0-100)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "remoteType",
            "required": false,
            "in": "query",
            "description": "Filtra per modalità di lavoro (REMOTE, HYBRID, ON_SITE)",
            "schema": {
              "enum": [
                "REMOTE",
                "HYBRID",
                "ON_SITE",
                "UNSPECIFIED"
              ],
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Cerca per testo nel titolo o nel nome azienda",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "required": false,
            "in": "query",
            "description": "Filtra per fonte dell annuncio (ARBEITNOW, REMOTIVE)",
            "schema": {
              "enum": [
                "ARBEITNOW",
                "REMOTIVE",
                "JOBICY",
                "WE_WORK_REMOTELY"
              ],
              "type": "string"
            }
          },
          {
            "name": "evaluatorModel",
            "required": false,
            "in": "query",
            "description": "Filtra per modello AI utilizzato per l analisi (es. GEMINI_3_1_FLASH_LITE, GEMMA_4_12B)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filtra per stato dell annuncio (es. NEW, SAVED, APPLIED, INTERVIEWING, REJECTED, ARCHIVED)",
            "schema": {
              "enum": [
                "NEW",
                "SAVED",
                "APPLIED",
                "SCREENING",
                "INTERVIEWING",
                "OFFER",
                "ACCEPTED",
                "REJECTED",
                "ARCHIVED"
              ],
              "type": "string"
            }
          },
          {
            "name": "includePending",
            "required": false,
            "in": "query",
            "description": "Includi anche gli annunci non ancora valutati dall AI (default: false)",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobOfferDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "job-offers"
        ]
      }
    },
    "/job-offers/closed": {
      "get": {
        "operationId": "JobOffersController_findClosed",
        "summary": "Recupera la lista degli annunci conclusi o scartati dall utente (REJECTED, ARCHIVED)",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "integer"
            }
          },
          {
            "name": "freshness",
            "required": false,
            "in": "query",
            "description": "Filtra per freschezza dell annuncio (HOT, RECENT, AGING, OLD)",
            "schema": {
              "enum": [
                "HOT",
                "RECENT",
                "AGING",
                "OLD"
              ],
              "type": "string"
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "description": "Filtra per priorità AI: HIGH, MEDIUM, LOW, DISQUALIFIED",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minDesireScore",
            "required": false,
            "in": "query",
            "description": "Filtra per punteggio minimo aderenza ai desiderata (0-100)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minCompetenceScore",
            "required": false,
            "in": "query",
            "description": "Filtra per punteggio minimo competenza tecnica (0-100)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minScore",
            "required": false,
            "in": "query",
            "description": "Filtra per punteggio globale minimo (0-100)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "remoteType",
            "required": false,
            "in": "query",
            "description": "Filtra per modalità di lavoro (REMOTE, HYBRID, ON_SITE)",
            "schema": {
              "enum": [
                "REMOTE",
                "HYBRID",
                "ON_SITE",
                "UNSPECIFIED"
              ],
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Cerca per testo nel titolo o nel nome azienda",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "required": false,
            "in": "query",
            "description": "Filtra per fonte dell annuncio (ARBEITNOW, REMOTIVE)",
            "schema": {
              "enum": [
                "ARBEITNOW",
                "REMOTIVE",
                "JOBICY",
                "WE_WORK_REMOTELY"
              ],
              "type": "string"
            }
          },
          {
            "name": "evaluatorModel",
            "required": false,
            "in": "query",
            "description": "Filtra per modello AI utilizzato per l analisi (es. GEMINI_3_1_FLASH_LITE, GEMMA_4_12B)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filtra per stato dell annuncio (es. NEW, SAVED, APPLIED, INTERVIEWING, REJECTED, ARCHIVED)",
            "schema": {
              "enum": [
                "NEW",
                "SAVED",
                "APPLIED",
                "SCREENING",
                "INTERVIEWING",
                "OFFER",
                "ACCEPTED",
                "REJECTED",
                "ARCHIVED"
              ],
              "type": "string"
            }
          },
          {
            "name": "includePending",
            "required": false,
            "in": "query",
            "description": "Includi anche gli annunci non ancora valutati dall AI (default: false)",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/JobOfferDto"
                          }
                        },
                        "meta": {
                          "$ref": "#/components/schemas/PaginatedMetaDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "job-offers"
        ]
      }
    },
    "/job-offers/disqualified": {
      "get": {
        "operationId": "JobOffersController_findDisqualified",
        "summary": "Recupera la lista degli annunci scartati dall AI (priorità DISQUALIFIED)",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "integer"
            }
          },
          {
            "name": "freshness",
            "required": false,
            "in": "query",
            "description": "Filtra per freschezza dell annuncio (HOT, RECENT, AGING, OLD)",
            "schema": {
              "enum": [
                "HOT",
                "RECENT",
                "AGING",
                "OLD"
              ],
              "type": "string"
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "description": "Filtra per priorità AI: HIGH, MEDIUM, LOW, DISQUALIFIED",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minDesireScore",
            "required": false,
            "in": "query",
            "description": "Filtra per punteggio minimo aderenza ai desiderata (0-100)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minCompetenceScore",
            "required": false,
            "in": "query",
            "description": "Filtra per punteggio minimo competenza tecnica (0-100)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minScore",
            "required": false,
            "in": "query",
            "description": "Filtra per punteggio globale minimo (0-100)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "remoteType",
            "required": false,
            "in": "query",
            "description": "Filtra per modalità di lavoro (REMOTE, HYBRID, ON_SITE)",
            "schema": {
              "enum": [
                "REMOTE",
                "HYBRID",
                "ON_SITE",
                "UNSPECIFIED"
              ],
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Cerca per testo nel titolo o nel nome azienda",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "required": false,
            "in": "query",
            "description": "Filtra per fonte dell annuncio (ARBEITNOW, REMOTIVE)",
            "schema": {
              "enum": [
                "ARBEITNOW",
                "REMOTIVE",
                "JOBICY",
                "WE_WORK_REMOTELY"
              ],
              "type": "string"
            }
          },
          {
            "name": "evaluatorModel",
            "required": false,
            "in": "query",
            "description": "Filtra per modello AI utilizzato per l analisi (es. GEMINI_3_1_FLASH_LITE, GEMMA_4_12B)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filtra per stato dell annuncio (es. NEW, SAVED, APPLIED, INTERVIEWING, REJECTED, ARCHIVED)",
            "schema": {
              "enum": [
                "NEW",
                "SAVED",
                "APPLIED",
                "SCREENING",
                "INTERVIEWING",
                "OFFER",
                "ACCEPTED",
                "REJECTED",
                "ARCHIVED"
              ],
              "type": "string"
            }
          },
          {
            "name": "includePending",
            "required": false,
            "in": "query",
            "description": "Includi anche gli annunci non ancora valutati dall AI (default: false)",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/JobOfferDto"
                          }
                        },
                        "meta": {
                          "$ref": "#/components/schemas/PaginatedMetaDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "job-offers"
        ]
      }
    },
...
    "/job-offers/{id}": {
      "get": {
        "operationId": "JobOffersController_findOne",
        "summary": "Recupera un singolo annuncio di lavoro tramite ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "job-offers"
        ]
      }
    },
    "/job-offers/{id}/curriculum/pdf": {
      "get": {
        "operationId": "JobOffersController_getCurriculumPdf",
        "summary": "Scarica o visualizza il file PDF del curriculum personalizzato generato per l annuncio",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "job-offers"
        ]
      }
    },
...
  },
  "info": {
    "title": "Analizzatore di Ruoli & Job Aggregator API",
    "description": "API Gateway per la ricerca, estrazione e gestione degli annunci di lavoro",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginatedMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
...
    }
  }
}